Add test for Qt Applications
This required the application to be split up into a library and two
executables, one for the actual application and the other for the tests.

QWidgets cannot be created if no `QApplication` exists, it is therefore
necessary to have a separate main function to create one before running
the tests.